+++ core/eventbus.gd
@@ -49,6 +49,8 @@ signal event_confirmed()
 
 signal request_change_left_panel_visibility(enable)
 signal request_toggle_map_only()
+signal request_toggle_debug_overlay()
+signal request_toggle_debug_controller()
 signal request_event(data: Variant, context: Dictionary)
 signal request_event_key(key: String, context: Dictionary)
 signal push_event(data: Variant, context: Dictionary)